rbind a rakelist

rbind a rakelist

rbind_rakelist(rakelist, doc_id = NULL)

Arguments

rakelist
An object of class rakelist, which you create by calling slowrake.

doc_id
An optional vector of document IDs, which should be the same length as rakelist. These IDs will be added to the resulting data frame.

Value

A single data frame which contains all documents' keywords. The doc_id column tells you which document a keyword was found in.

Examples

rakelist <- slowrake(txt = dog_pubs$abstract[1:2])

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%


# Without specifying doc_id:
rbind_rakelist(rakelist = rakelist)

#>     doc_id                                   keyword freq     score
#> 1        1        assistance dog identification tags    1 10.833333
#> 2        1                 animal control facilities    1  9.000000
#> 3        1                 emotional support animals    1  9.000000
#> 4        1                          small body sizes    1  9.000000
#> 5        1              seemingly inappropriate dogs    1  7.916667
#> 6        1             governmental oversight system    1  7.333333
#> 7        1                   assistance dogs sharply    1  7.333333
#> 8        1                registered assistance dogs    1  7.333333
#> 9        1               unqualified assistance dogs    1  7.333333
#> 10       1                        medical assistance    1  4.416667
#> 11       1                  governmental enforcement    1  4.333333
#> 12       1                            assistance dog    1  4.333333
#> 13       1                           assistance dogs    6  4.333333
#> 14       1                         chaotic situation    1  4.000000
#> 15       1                         management system    1  4.000000
#> 16       1                               past decade    1  4.000000
#> 17       1                             public access    1  4.000000
#> 18       1                                 dog bites    1  3.916667
#> 19       1                              service dogs    3  3.916667
#> 20       1                   california governmental    1  3.583333
#> 21       1                      current demographics    1  3.500000
#> 22       1                              federal laws    1  3.500000
#> 23       1                            identification    1  2.500000
#> 24       1                                    system    1  2.000000
#> 25       1                                       dog    1  1.916667
#> 26       1                                      dogs    7  1.916667
#> 27       1                                   current    1  1.500000
#> 28       1                                       law    1  1.500000
#> 29       1                                california    3  1.250000
#> 30       1                                adequately    1  1.000000
#> 31       1                                       age    1  1.000000
#> 32       1                                      cats    1  1.000000
#> 33       1                             certification    1  1.000000
#> 34       1                                      data    1  1.000000
#> 35       1                                definition    1  1.000000
#> 36       1                              disabilities    2  1.000000
#> 37       1                                    errors    1  1.000000
#> 38       1                                  handlers    1  1.000000
#> 39       1                                 incidents    1  1.000000
#> 40       1                                  increase    1  1.000000
#> 41       1                               information    1  1.000000
#> 42       1                               legislation    1  1.000000
#> 43       1                                mistakenly    1  1.000000
#> 44       1                          misunderstanding    1  1.000000
#> 45       1                                    misuse    1  1.000000
#> 46       1                                    number    2  1.000000
#> 47       1                                     older    1  1.000000
#> 48       1                                  patterns    1  1.000000
#> 49       1                                    people    2  1.000000
#> 50       1                                prevalence    1  1.000000
#> 51       1                                  problems    2  1.000000
#> 52       1                               psychiatric    1  1.000000
#> 53       1                               regulations    1  1.000000
#> 54       1                                     roles    1  1.000000
#> 55       1                                statistics    1  1.000000
#> 56       1                                  strongly    1  1.000000
#> 57       1                                     study    1  1.000000
#> 58       1                               substantive    1  1.000000
#> 59       1                                      time    1  1.000000
#> 60       1                                     types    1  1.000000
#> 61       1                                     years    1  1.000000
#> 62       2 current dog suitability assessments focus    1 21.263158
#> 63       2                     guide dog assessments    1  8.808612
#> 64       2              potential multifarious roles    1  8.000000
#> 65       2              structured interview surveys    1  8.000000
#> 66       2                           guide dog owner    1  6.697501
#> 67       2                          guide dog owners    2  6.697501
#> 68       2                       guide dog behaviour    2  6.451470
#> 69       2                 important positive aspect    1  6.083333
#> 70       2                                 guide dog    5  4.808612
#> 71       2                           important areas    1  4.333333
#> 72       2                            key importance    1  4.333333
#> 73       2                              domestic dog    1  4.263158
#> 74       2                                dog owners    1  4.152047
#> 75       2                       animal interactions    1  4.000000
#> 76       2                          canis familiaris    1  4.000000
#> 77       2                           decision making    1  4.000000
#> 78       2                             energy levels    1  4.000000
#> 79       2                     environmental stimuli    1  4.000000
#> 80       2                              high tension    1  4.000000
#> 81       2                                 key areas    1  4.000000
#> 82       2                               large range    1  4.000000
#> 83       2                        low aggressiveness    1  4.000000
#> 84       2                          mobility support    1  4.000000
#> 85       2                         thematic analysis    1  4.000000
#> 86       2                      typical expectations    1  4.000000
#> 87       2                            confident dogs    1  3.929825
#> 88       2                           negative aspect    1  3.750000
#> 89       2                      attachment behaviour    1  3.642857
#> 90       2                        behavioural traits    1  3.642857
#> 91       2                            safe behaviour    1  3.642857
#> 92       2                          stress behaviour    1  3.642857
#> 93       2                          confident owners    1  3.555556
#> 94       2                         negative examples    1  3.500000
#> 95       2                       pleasing behaviours    1  3.142857
#> 96       2                                       dog    3  2.263158
#> 97       2                                      dogs    1  2.263158
#> 98       2                                  positive    1  2.000000
#> 99       2                                      role    1  2.000000
#> 100      2                                    survey    1  2.000000
#> 101      2                                     owner    2  1.888889
#> 102      2                                    owners    2  1.888889
#> 103      2                                   aspects    2  1.750000
#> 104      2                                confidence    1  1.666667
#> 105      2                                 behaviour    7  1.642857
#> 106      2                                  examples    1  1.500000
#> 107      2                                  pleasing    1  1.500000
#> 108      2                                       aim    1  1.000000
#> 109      2                                 attention    1  1.000000
#> 110      2                             attentiveness    1  1.000000
#> 111      2                                     blind    1  1.000000
#> 112      2                                  calmness    1  1.000000
#> 113      2                                  children    1  1.000000
#> 114      2                               consistency    1  1.000000
#> 115      2                                   control    1  1.000000
#> 116      2                             disappointing    1  1.000000
#> 117      2                                 discovery    1  1.000000
#> 118      2                               environment    1  1.000000
#> 119      2                                      face    1  1.000000
#> 120      2                               fearfulness    1  1.000000
#> 121      2                                   harness    1  1.000000
#> 122      2                             heterogeneity    1  1.000000
#> 123      2                               highlighted    1  1.000000
#> 124      2                                highlights    1  1.000000
#> 125      2                                     human    1  1.000000
#> 126      2                             inappropriate    1  1.000000
#> 127      2                                      lead    1  1.000000
#> 128      2                                literature    1  1.000000
#> 129      2                                    living    1  1.000000
#> 130      2                                  maturity    1  1.000000
#> 131      2                                    nature    1  1.000000
#> 132      2                                 obedience    1  1.000000
#> 133      2                                objectives    1  1.000000
#> 134      2                              perspectives    1  1.000000
#> 135      2                                priorities    1  1.000000
#> 136      2                               problematic    1  1.000000
#> 137      2                                reactivity    1  1.000000
#> 138      2                                  relation    1  1.000000
#> 139      2                                   results    2  1.000000
#> 140      2                                      semi    1  1.000000
#> 141      2                                situations    1  1.000000
#> 142      2                                     sixty    1  1.000000
#> 143      2                              specifically    1  1.000000
#> 144      2                                     study    2  1.000000
#> 145      2                                      task    1  1.000000
#> 146      2                                    themes    1  1.000000
#> 147      2                                    topics    1  1.000000
#> 148      2                                   traffic    1  1.000000
#> 149      2                              trainability    1  1.000000
#> 150      2                                  visually    1  1.000000
#> 151      2                                      work    2  1.000000
#>                                stem
#> 1            assist dog identif tag
#> 2                anim control facil
#> 3                 emot support anim
#> 4                   small bodi size
#> 5          seemingli inappropri dog
#> 6       government oversight system
#> 7                assist dog sharpli
#> 8                 regist assist dog
#> 9              unqualifi assist dog
#> 10                     medic assist
#> 11                government enforc
#> 12                       assist dog
#> 13                       assist dog
#> 14                   chaotic situat
#> 15                     manag system
#> 16                       past decad
#> 17                    public access
#> 18                         dog bite
#> 19                       servic dog
#> 20            california government
#> 21                current demograph
#> 22                        feder law
#> 23                          identif
#> 24                           system
#> 25                              dog
#> 26                              dog
#> 27                          current
#> 28                              law
#> 29                       california
#> 30                            adequ
#> 31                               ag
#> 32                              cat
#> 33                           certif
#> 34                             data
#> 35                          definit
#> 36                           disabl
#> 37                            error
#> 38                          handler
#> 39                            incid
#> 40                          increas
#> 41                           inform
#> 42                           legisl
#> 43                       mistakenli
#> 44                    misunderstand
#> 45                            misus
#> 46                           number
#> 47                            older
#> 48                          pattern
#> 49                            peopl
#> 50                           preval
#> 51                          problem
#> 52                        psychiatr
#> 53                            regul
#> 54                             role
#> 55                          statist
#> 56                         strongli
#> 57                            studi
#> 58                         substant
#> 59                             time
#> 60                             type
#> 61                             year
#> 62  current dog suitabl assess focu
#> 63                  guid dog assess
#> 64           potenti multifari role
#> 65        structur interview survei
#> 66                   guid dog owner
#> 67                   guid dog owner
#> 68               guid dog behaviour
#> 69              import posit aspect
#> 70                         guid dog
#> 71                      import area
#> 72                       kei import
#> 73                       domest dog
#> 74                        dog owner
#> 75                    anim interact
#> 76                   cani familiari
#> 77                       decis make
#> 78                     energi level
#> 79              environment stimuli
#> 80                     high tension
#> 81                         kei area
#> 82                        larg rang
#> 83                      low aggress
#> 84                    mobil support
#> 85                   themat analysi
#> 86                     typic expect
#> 87                       confid dog
#> 88                       neg aspect
#> 89                 attach behaviour
#> 90                  behaviour trait
#> 91                   safe behaviour
#> 92                 stress behaviour
#> 93                     confid owner
#> 94                       neg exampl
#> 95                  pleas behaviour
#> 96                              dog
#> 97                              dog
#> 98                            posit
#> 99                             role
#> 100                          survei
#> 101                           owner
#> 102                           owner
#> 103                          aspect
#> 104                          confid
#> 105                       behaviour
#> 106                          exampl
#> 107                           pleas
#> 108                             aim
#> 109                          attent
#> 110                          attent
#> 111                           blind
#> 112                            calm
#> 113                        children
#> 114                         consist
#> 115                         control
#> 116                      disappoint
#> 117                       discoveri
#> 118                         environ
#> 119                            face
#> 120                            fear
#> 121                             har
#> 122                       heterogen
#> 123                       highlight
#> 124                       highlight
#> 125                           human
#> 126                      inappropri
#> 127                            lead
#> 128                       literatur
#> 129                            live
#> 130                           matur
#> 131                           natur
#> 132                           obedi
#> 133                          object
#> 134                        perspect
#> 135                        prioriti
#> 136                       problemat
#> 137                         reactiv
#> 138                           relat
#> 139                          result
#> 140                            semi
#> 141                          situat
#> 142                           sixti
#> 143                          specif
#> 144                           studi
#> 145                            task
#> 146                           theme
#> 147                           topic
#> 148                         traffic
#> 149                        trainabl
#> 150                          visual
#> 151                            work


# With specifying doc_id:
rbind_rakelist(rakelist = rakelist, doc_id = dog_pubs$doi[1:2])

#>                           doc_id                                   keyword freq
#> 1   10.1371/journal.pone.0132820        assistance dog identification tags    1
#> 2   10.1371/journal.pone.0132820                 animal control facilities    1
#> 3   10.1371/journal.pone.0132820                 emotional support animals    1
#> 4   10.1371/journal.pone.0132820                          small body sizes    1
#> 5   10.1371/journal.pone.0132820              seemingly inappropriate dogs    1
#> 6   10.1371/journal.pone.0132820             governmental oversight system    1
#> 7   10.1371/journal.pone.0132820                   assistance dogs sharply    1
#> 8   10.1371/journal.pone.0132820                registered assistance dogs    1
#> 9   10.1371/journal.pone.0132820               unqualified assistance dogs    1
#> 10  10.1371/journal.pone.0132820                        medical assistance    1
#> 11  10.1371/journal.pone.0132820                  governmental enforcement    1
#> 12  10.1371/journal.pone.0132820                            assistance dog    1
#> 13  10.1371/journal.pone.0132820                           assistance dogs    6
#> 14  10.1371/journal.pone.0132820                         chaotic situation    1
#> 15  10.1371/journal.pone.0132820                         management system    1
#> 16  10.1371/journal.pone.0132820                               past decade    1
#> 17  10.1371/journal.pone.0132820                             public access    1
#> 18  10.1371/journal.pone.0132820                                 dog bites    1
#> 19  10.1371/journal.pone.0132820                              service dogs    3
#> 20  10.1371/journal.pone.0132820                   california governmental    1
#> 21  10.1371/journal.pone.0132820                      current demographics    1
#> 22  10.1371/journal.pone.0132820                              federal laws    1
#> 23  10.1371/journal.pone.0132820                            identification    1
#> 24  10.1371/journal.pone.0132820                                    system    1
#> 25  10.1371/journal.pone.0132820                                       dog    1
#> 26  10.1371/journal.pone.0132820                                      dogs    7
#> 27  10.1371/journal.pone.0132820                                   current    1
#> 28  10.1371/journal.pone.0132820                                       law    1
#> 29  10.1371/journal.pone.0132820                                california    3
#> 30  10.1371/journal.pone.0132820                                adequately    1
#> 31  10.1371/journal.pone.0132820                                       age    1
#> 32  10.1371/journal.pone.0132820                                      cats    1
#> 33  10.1371/journal.pone.0132820                             certification    1
#> 34  10.1371/journal.pone.0132820                                      data    1
#> 35  10.1371/journal.pone.0132820                                definition    1
#> 36  10.1371/journal.pone.0132820                              disabilities    2
#> 37  10.1371/journal.pone.0132820                                    errors    1
#> 38  10.1371/journal.pone.0132820                                  handlers    1
#> 39  10.1371/journal.pone.0132820                                 incidents    1
#> 40  10.1371/journal.pone.0132820                                  increase    1
#> 41  10.1371/journal.pone.0132820                               information    1
#> 42  10.1371/journal.pone.0132820                               legislation    1
#> 43  10.1371/journal.pone.0132820                                mistakenly    1
#> 44  10.1371/journal.pone.0132820                          misunderstanding    1
#> 45  10.1371/journal.pone.0132820                                    misuse    1
#> 46  10.1371/journal.pone.0132820                                    number    2
#> 47  10.1371/journal.pone.0132820                                     older    1
#> 48  10.1371/journal.pone.0132820                                  patterns    1
#> 49  10.1371/journal.pone.0132820                                    people    2
#> 50  10.1371/journal.pone.0132820                                prevalence    1
#> 51  10.1371/journal.pone.0132820                                  problems    2
#> 52  10.1371/journal.pone.0132820                               psychiatric    1
#> 53  10.1371/journal.pone.0132820                               regulations    1
#> 54  10.1371/journal.pone.0132820                                     roles    1
#> 55  10.1371/journal.pone.0132820                                statistics    1
#> 56  10.1371/journal.pone.0132820                                  strongly    1
#> 57  10.1371/journal.pone.0132820                                     study    1
#> 58  10.1371/journal.pone.0132820                               substantive    1
#> 59  10.1371/journal.pone.0132820                                      time    1
#> 60  10.1371/journal.pone.0132820                                     types    1
#> 61  10.1371/journal.pone.0132820                                     years    1
#> 62  10.1371/journal.pone.0176018 current dog suitability assessments focus    1
#> 63  10.1371/journal.pone.0176018                     guide dog assessments    1
#> 64  10.1371/journal.pone.0176018              potential multifarious roles    1
#> 65  10.1371/journal.pone.0176018              structured interview surveys    1
#> 66  10.1371/journal.pone.0176018                           guide dog owner    1
#> 67  10.1371/journal.pone.0176018                          guide dog owners    2
#> 68  10.1371/journal.pone.0176018                       guide dog behaviour    2
#> 69  10.1371/journal.pone.0176018                 important positive aspect    1
#> 70  10.1371/journal.pone.0176018                                 guide dog    5
#> 71  10.1371/journal.pone.0176018                           important areas    1
#> 72  10.1371/journal.pone.0176018                            key importance    1
#> 73  10.1371/journal.pone.0176018                              domestic dog    1
#> 74  10.1371/journal.pone.0176018                                dog owners    1
#> 75  10.1371/journal.pone.0176018                       animal interactions    1
#> 76  10.1371/journal.pone.0176018                          canis familiaris    1
#> 77  10.1371/journal.pone.0176018                           decision making    1
#> 78  10.1371/journal.pone.0176018                             energy levels    1
#> 79  10.1371/journal.pone.0176018                     environmental stimuli    1
#> 80  10.1371/journal.pone.0176018                              high tension    1
#> 81  10.1371/journal.pone.0176018                                 key areas    1
#> 82  10.1371/journal.pone.0176018                               large range    1
#> 83  10.1371/journal.pone.0176018                        low aggressiveness    1
#> 84  10.1371/journal.pone.0176018                          mobility support    1
#> 85  10.1371/journal.pone.0176018                         thematic analysis    1
#> 86  10.1371/journal.pone.0176018                      typical expectations    1
#> 87  10.1371/journal.pone.0176018                            confident dogs    1
#> 88  10.1371/journal.pone.0176018                           negative aspect    1
#> 89  10.1371/journal.pone.0176018                      attachment behaviour    1
#> 90  10.1371/journal.pone.0176018                        behavioural traits    1
#> 91  10.1371/journal.pone.0176018                            safe behaviour    1
#> 92  10.1371/journal.pone.0176018                          stress behaviour    1
#> 93  10.1371/journal.pone.0176018                          confident owners    1
#> 94  10.1371/journal.pone.0176018                         negative examples    1
#> 95  10.1371/journal.pone.0176018                       pleasing behaviours    1
#> 96  10.1371/journal.pone.0176018                                       dog    3
#> 97  10.1371/journal.pone.0176018                                      dogs    1
#> 98  10.1371/journal.pone.0176018                                  positive    1
#> 99  10.1371/journal.pone.0176018                                      role    1
#> 100 10.1371/journal.pone.0176018                                    survey    1
#> 101 10.1371/journal.pone.0176018                                     owner    2
#> 102 10.1371/journal.pone.0176018                                    owners    2
#> 103 10.1371/journal.pone.0176018                                   aspects    2
#> 104 10.1371/journal.pone.0176018                                confidence    1
#> 105 10.1371/journal.pone.0176018                                 behaviour    7
#> 106 10.1371/journal.pone.0176018                                  examples    1
#> 107 10.1371/journal.pone.0176018                                  pleasing    1
#> 108 10.1371/journal.pone.0176018                                       aim    1
#> 109 10.1371/journal.pone.0176018                                 attention    1
#> 110 10.1371/journal.pone.0176018                             attentiveness    1
#> 111 10.1371/journal.pone.0176018                                     blind    1
#> 112 10.1371/journal.pone.0176018                                  calmness    1
#> 113 10.1371/journal.pone.0176018                                  children    1
#> 114 10.1371/journal.pone.0176018                               consistency    1
#> 115 10.1371/journal.pone.0176018                                   control    1
#> 116 10.1371/journal.pone.0176018                             disappointing    1
#> 117 10.1371/journal.pone.0176018                                 discovery    1
#> 118 10.1371/journal.pone.0176018                               environment    1
#> 119 10.1371/journal.pone.0176018                                      face    1
#> 120 10.1371/journal.pone.0176018                               fearfulness    1
#> 121 10.1371/journal.pone.0176018                                   harness    1
#> 122 10.1371/journal.pone.0176018                             heterogeneity    1
#> 123 10.1371/journal.pone.0176018                               highlighted    1
#> 124 10.1371/journal.pone.0176018                                highlights    1
#> 125 10.1371/journal.pone.0176018                                     human    1
#> 126 10.1371/journal.pone.0176018                             inappropriate    1
#> 127 10.1371/journal.pone.0176018                                      lead    1
#> 128 10.1371/journal.pone.0176018                                literature    1
#> 129 10.1371/journal.pone.0176018                                    living    1
#> 130 10.1371/journal.pone.0176018                                  maturity    1
#> 131 10.1371/journal.pone.0176018                                    nature    1
#> 132 10.1371/journal.pone.0176018                                 obedience    1
#> 133 10.1371/journal.pone.0176018                                objectives    1
#> 134 10.1371/journal.pone.0176018                              perspectives    1
#> 135 10.1371/journal.pone.0176018                                priorities    1
#> 136 10.1371/journal.pone.0176018                               problematic    1
#> 137 10.1371/journal.pone.0176018                                reactivity    1
#> 138 10.1371/journal.pone.0176018                                  relation    1
#> 139 10.1371/journal.pone.0176018                                   results    2
#> 140 10.1371/journal.pone.0176018                                      semi    1
#> 141 10.1371/journal.pone.0176018                                situations    1
#> 142 10.1371/journal.pone.0176018                                     sixty    1
#> 143 10.1371/journal.pone.0176018                              specifically    1
#> 144 10.1371/journal.pone.0176018                                     study    2
#> 145 10.1371/journal.pone.0176018                                      task    1
#> 146 10.1371/journal.pone.0176018                                    themes    1
#> 147 10.1371/journal.pone.0176018                                    topics    1
#> 148 10.1371/journal.pone.0176018                                   traffic    1
#> 149 10.1371/journal.pone.0176018                              trainability    1
#> 150 10.1371/journal.pone.0176018                                  visually    1
#> 151 10.1371/journal.pone.0176018                                      work    2
#>         score                            stem
#> 1   10.833333          assist dog identif tag
#> 2    9.000000              anim control facil
#> 3    9.000000               emot support anim
#> 4    9.000000                 small bodi size
#> 5    7.916667        seemingli inappropri dog
#> 6    7.333333     government oversight system
#> 7    7.333333              assist dog sharpli
#> 8    7.333333               regist assist dog
#> 9    7.333333            unqualifi assist dog
#> 10   4.416667                    medic assist
#> 11   4.333333               government enforc
#> 12   4.333333                      assist dog
#> 13   4.333333                      assist dog
#> 14   4.000000                  chaotic situat
#> 15   4.000000                    manag system
#> 16   4.000000                      past decad
#> 17   4.000000                   public access
#> 18   3.916667                        dog bite
#> 19   3.916667                      servic dog
#> 20   3.583333           california government
#> 21   3.500000               current demograph
#> 22   3.500000                       feder law
#> 23   2.500000                         identif
#> 24   2.000000                          system
#> 25   1.916667                             dog
#> 26   1.916667                             dog
#> 27   1.500000                         current
#> 28   1.500000                             law
#> 29   1.250000                      california
#> 30   1.000000                           adequ
#> 31   1.000000                              ag
#> 32   1.000000                             cat
#> 33   1.000000                          certif
#> 34   1.000000                            data
#> 35   1.000000                         definit
#> 36   1.000000                          disabl
#> 37   1.000000                           error
#> 38   1.000000                         handler
#> 39   1.000000                           incid
#> 40   1.000000                         increas
#> 41   1.000000                          inform
#> 42   1.000000                          legisl
#> 43   1.000000                      mistakenli
#> 44   1.000000                   misunderstand
#> 45   1.000000                           misus
#> 46   1.000000                          number
#> 47   1.000000                           older
#> 48   1.000000                         pattern
#> 49   1.000000                           peopl
#> 50   1.000000                          preval
#> 51   1.000000                         problem
#> 52   1.000000                       psychiatr
#> 53   1.000000                           regul
#> 54   1.000000                            role
#> 55   1.000000                         statist
#> 56   1.000000                        strongli
#> 57   1.000000                           studi
#> 58   1.000000                        substant
#> 59   1.000000                            time
#> 60   1.000000                            type
#> 61   1.000000                            year
#> 62  21.263158 current dog suitabl assess focu
#> 63   8.808612                 guid dog assess
#> 64   8.000000          potenti multifari role
#> 65   8.000000       structur interview survei
#> 66   6.697501                  guid dog owner
#> 67   6.697501                  guid dog owner
#> 68   6.451470              guid dog behaviour
#> 69   6.083333             import posit aspect
#> 70   4.808612                        guid dog
#> 71   4.333333                     import area
#> 72   4.333333                      kei import
#> 73   4.263158                      domest dog
#> 74   4.152047                       dog owner
#> 75   4.000000                   anim interact
#> 76   4.000000                  cani familiari
#> 77   4.000000                      decis make
#> 78   4.000000                    energi level
#> 79   4.000000             environment stimuli
#> 80   4.000000                    high tension
#> 81   4.000000                        kei area
#> 82   4.000000                       larg rang
#> 83   4.000000                     low aggress
#> 84   4.000000                   mobil support
#> 85   4.000000                  themat analysi
#> 86   4.000000                    typic expect
#> 87   3.929825                      confid dog
#> 88   3.750000                      neg aspect
#> 89   3.642857                attach behaviour
#> 90   3.642857                 behaviour trait
#> 91   3.642857                  safe behaviour
#> 92   3.642857                stress behaviour
#> 93   3.555556                    confid owner
#> 94   3.500000                      neg exampl
#> 95   3.142857                 pleas behaviour
#> 96   2.263158                             dog
#> 97   2.263158                             dog
#> 98   2.000000                           posit
#> 99   2.000000                            role
#> 100  2.000000                          survei
#> 101  1.888889                           owner
#> 102  1.888889                           owner
#> 103  1.750000                          aspect
#> 104  1.666667                          confid
#> 105  1.642857                       behaviour
#> 106  1.500000                          exampl
#> 107  1.500000                           pleas
#> 108  1.000000                             aim
#> 109  1.000000                          attent
#> 110  1.000000                          attent
#> 111  1.000000                           blind
#> 112  1.000000                            calm
#> 113  1.000000                        children
#> 114  1.000000                         consist
#> 115  1.000000                         control
#> 116  1.000000                      disappoint
#> 117  1.000000                       discoveri
#> 118  1.000000                         environ
#> 119  1.000000                            face
#> 120  1.000000                            fear
#> 121  1.000000                             har
#> 122  1.000000                       heterogen
#> 123  1.000000                       highlight
#> 124  1.000000                       highlight
#> 125  1.000000                           human
#> 126  1.000000                      inappropri
#> 127  1.000000                            lead
#> 128  1.000000                       literatur
#> 129  1.000000                            live
#> 130  1.000000                           matur
#> 131  1.000000                           natur
#> 132  1.000000                           obedi
#> 133  1.000000                          object
#> 134  1.000000                        perspect
#> 135  1.000000                        prioriti
#> 136  1.000000                       problemat
#> 137  1.000000                         reactiv
#> 138  1.000000                           relat
#> 139  1.000000                          result
#> 140  1.000000                            semi
#> 141  1.000000                          situat
#> 142  1.000000                           sixti
#> 143  1.000000                          specif
#> 144  1.000000                           studi
#> 145  1.000000                            task
#> 146  1.000000                           theme
#> 147  1.000000                           topic
#> 148  1.000000                         traffic
#> 149  1.000000                        trainabl
#> 150  1.000000                          visual
#> 151  1.000000                            work